QuestionAI/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.27.0
redis==4.5.5
fastapi-limiter==0.1.5
python-jose==3.3.0
passlib==1.7.4
